Moving method to case data
